Refactor fragment locations
